perf(home): unchanged wallet refreshes no longer rebuild the transaction list (#1014)
Four compounding causes made every Home wallet churn re-render all
visible transaction rows (measured: 11 commits of 103-158 ms across 3
pull-to-refreshes on a near-empty account; 111 ms of every 140 ms Home
commit was the list):

1. fetchTransactionsForWallet committed a rebuilt-but-equal transactions
   array on every poll. Now reuses the JSON it already computes for the
   AsyncStorage persist as a content fingerprint and keeps the old array
   identity (and skips the redundant disk write) when unchanged.
2. TransactionList's swapStateTick (the FlatList extraData) bumped
   unconditionally on every swap-recovery notify - and pull-to-refresh
   always kicks a recovery pass. Now fingerprints the attention/claimed
   sets and bumps only on real change.
3. mergeResolverResults wrote zapCounterparty: null over rows that
   already had no attribution, rebuilding the array as a no-op. Now
   leaves untouched rows (and the array identity) alone.
4. FlatList re-wraps renderItem whenever it re-renders (and the fresh
   RefreshControl element defeated its PureComponent bail every Home
   commit), re-executing every visible cell. Row content is now the
   memoised TransactionRow (the #703 ConversationMessageRow pattern) so
   a re-executed cell bails on stable props, and Home memoises the
   RefreshControl element.

After: zero list commits >=100 ms in the same scenario; Home commits
23-35 ms (was 103-158 ms). NWC connection watchdog extracted to
useNwcConnectionWatchdog.ts to keep WalletContext shrinking (2173 ->
2153 lines; gate baseline updated).
