+++ scripts/check-file-size.sh
@@ -17,7 +17,7 @@ LIMIT=1000
 declare -A BASELINE=(
   ["src/contexts/NostrContext.tsx"]=1837
   ["src/screens/HuntCreateScreen.tsx"]=2386
-  ["src/contexts/WalletContext.tsx"]=2173
+  ["src/contexts/WalletContext.tsx"]=2153
   ["src/screens/HuntPiggyDetailScreen.tsx"]=1710
   ["src/screens/MapScreen.tsx"]=1562
   ["src/services/nostrService.ts"]=1475
